api/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Async support
httpx==0.25.2
asyncio==3.4.3
aiohttp==3.9.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.8.0
pre-commit==3.6.0

# Additional utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Redis for rate limiting
redis==5.0.1
hiredis==2.3.2

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.7

# Google Cloud Platform dependencies
google-cloud-firestore==2.14.0
google-cloud-tasks==2.15.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.13.0
google-cloud-monitoring==2.18.0
google-cloud-trace==1.12.0
google-cloud-error-reporting==1.10.0

# OpenTelemetry for distributed tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-exporter-gcp-monitoring==1.6.0a0

# Additional async support for GCP
aiofiles==23.2.1
